backend/requirements.txt
Django==5.0.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.3.1
python-dotenv==1.0.1
opencv-python==4.9.0.80
numpy==1.26.4
pytesseract==0.3.10
Pillow==10.2.0
requests==2.31.0
channels channels-redis
pillow 
gunicorn  
whitenoise 
dj-database-url  
 sentry-sdk  
python-magic  